requirements.txt
PyQt5~=5.15.9
Pillow~=9.5.0
docx~=0.2.4